build: allow WiFi AP customization during build
This commit adds a sub-menu to the image options to modify the default
access point (AP) automatically created on first boot. The OpenWrt WiFi
AP is traditionally disabled for security reasons and does not require a
password (encryption "none").

The package base-files will generate a file called
`/etc/openwrt_wifi_defaults` which is read by the mac80211.sh script and
allows modification of the UCI options `disabled`, `ssid`, `encryption`
and `key`.

More sophisticated setups are not supported this is merely used to allow
modification of the default AP, just like `VERSIONOPT` is used to modify
the `/etc/openwrt_release` file.

Signed-off-by: Paul Spooren <mail@aparcar.org>